caps/tests/mochitest/test_loadLocalFont.html

<!DOCTYPE HTML>
<html>
<head>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<script type="application/javascript">

try {
  let Ahem = new FontFace(
    "Ahem",
    "url('chrome://mochitests/content/tests/dom/base/test/Ahem.ttf')"
  ).load().then(() => {
    ok(true, "We should succeed (resolve)");
  }, () => {
    ok(false, "We shouldn't fail (reject)");
  });
  console.log(Ahem); // avoid warning
} catch (e) {
  ok(false, "We shouldn't fail (catch)" + e);
}

</script>
</pre>
</body>
</html>
